Integrate i2c and compile

+++ i2c.h
@@ -13,7 +13,7 @@
 #include <stdint.h>
 
 void i2c_write(uint8_t addr, uint8_t data);
-uint8_t i2c_read(uint8_t addr);
+unsigned int i2c_read(uint8_t addr);
 
 
 #endif /* I2C_H_ */
